+++ CHANGELOG.md
@@ -8,6 +8,7 @@ We use [semantic versioning](http://semver.org/):
 - [breaking] _teamscale-gradle-plugin_: Removed usage of deprecated Gradle APIs.
   - `teamscale.agent.destinationProperty` type has changed from `Property<File>` to `DirectoryProperty`
   - `teamscale.agent.setDestination(destination)` parameter type has changed from `Property<File>` to `Provider<Directory>`
+- [fix] _agent_: Discovery of unsupported class file versions will no longer lead to a crash, but will be logged instead
 
 # 31.0.0
 - [breaking] Replaced `teamscale-git-properties-jar` with `git-properties-jar`. Jars/Wars/Ears/Aars provided with this option will now also be searched recursively for git.properties files except you set `search-git-properties-recursively=false`.